gdb/testsuite: make DWARF assembler's ranges' "base" and "range" procs
When creating a .debug_ranges section using the testsuite's DWARF
assembler, it currently looks like this:

  ranges {
    sequence {
      {base ...}
      {range ...}
      {range ...}
    }
  }

The sub-tree of sequence is manually traversed as a list of lists.  I
think it would be nicer if `base` and `range` where procedure, just like
the other levels:

  ranges {
    sequence {
      base ...
      range ...
      range ...
    }
  }

That makes the implementation more robust, and the usage a bit nicer
(less special characters).  It also allows having comments in between
the range list entries:

  ranges {
    sequence {
      base ...
      range ...

      # Hello world.
      range ...
    }
  }

... which doesn't work with the current approach.

gdb/testsuite/ChangeLog:

	* lib/dwarf.exp (ranges): Handle "base" and "range" as
	proceduresu.
	* gdb.dwarf/dw2-bad-elf.exp: Adjust.
	* gdb.dwarf2/dw2-inline-many-frames.exp: Adjust.
	* gdb.dwarf2/dw2-inline-stepping.exp: Adjust.
	* gdb.dwarf2/dw2-ranges-base.exp: Adjust.
	* gdb.dwarf2/dw2-ranges-func.exp: Adjust.
	* gdb.dwarf2/dw2-ranges-overlap.exp: Adjust.
	* gdb.dwarf2/dw2-ranges-psym.exp: Adjust.
	* gdb.dwarf2/enqueued-cu-base-addr.exp: Adjust.

Change-Id: I0b2af480faff54d0fd4214e0cc8d042d9583a865